Muutoksia logiikassa DB:n kanssa, määrittelen lisää, kun saan valmiiksi, mutta tässä vielä vähän keskeneräinen tuotos joka jokseenkin toimii.

+++ app/src/main/java/com/example/elecstore/RealtimeDatabase.kt
@@ -19,11 +19,11 @@ class RealtimeDatabase {
 
     var prodnum = 0
     var prodname = ""
-    var prodprice = 0
+    var prodprice = 0.0
     var prodval = 0
 
     constructor()
-    constructor(prodname: String, prodprice: Int, prodnum: Int, prodval: Int)
+    constructor(prodname: String, prodprice: Double, prodnum: Int, prodval: Int)
     {
         this.prodnum = prodnum
         this.prodprice = prodprice
